+ Improved AI_Base and AI_WithAttackBehavior. All AI's now have a new InteractWithTarget method that is called by AI_Base when it is at and should interact with a non-combat target. Timing of interactions is handled automatically just like with attacks in AI_BaseWithAttackBehavior. + AIs now have a new IsInteracting property + AIs behave better when they get in each others' way now + Added new building: bridge + Expanded the test map so there is a canyon to build bridges on to provide access to new terrain so the player can get more room to expand.
NOTE: Bridges are not fully implemented. Its construction needs to be limited to areas where it makes sense to build a bridge while not allowing other buildings to be constructed in such places. ALL building models and level models must now have Read/Write enabled in the build settings, because the foundation for making the game auto-regenerate the navmeshes is now in place as well. It will need to do this any time a bridge is constructed or destroyed, but this is not implemented just yet.